refactor(templates): move Template ORM into app/templates/models.py (Resolves #255)
## Summary
- Relocate the `Template` SQLAlchemy ORM class from
  `app/servers/models.py` to a new `app/templates/models.py`, ending
  the templates-table co-location with the servers domain.
- Keep `Server.template = relationship("Template", ...)` declared on
  the `Server` model; SQLAlchemy resolves the string-based class
  reference at mapper-configuration time once the `Template` module
  has been imported.
- Update three callsites that imported `Template` from
  `app.servers.models` (repository adapter + two test modules).
- `__tablename__` is unchanged (`templates`), so no Alembic / schema
  migration is required and existing rows are preserved.

## Mapper-config order guarantee
`app/templates/__init__.py` now eagerly imports `models`
(`from . import models  # noqa: F401`). Any code that imports the
`app.templates` package — including `from app.templates.router import
router` in `app.main` — guarantees the `Template` class is registered
with `Base.metadata` before `Base.metadata.create_all()` runs in the
startup lifespan, and before the `Server.template` back-population
resolves at first mapper configuration. The pre-existing chain
`templates.router → application → adapters.repository → templates.models`
provides belt-and-braces coverage.

## Verification
- `rg -n 'from app.servers.models import .*Template' app/ tests/` -> 0 hits
- `uv run ruff check` on touched files: clean
- `uv run pytest tests/unit -m "not slow"`: pass
- `uv run pytest tests/integration -m "not slow"`: pass
- `just test` (full suite incl. slow): 1718 passed, 10 skipped
- `uv run pre-commit run --files <touched>`: pass
